Add PostgreSQL and MongoDB services to docker-compose.yml
This update introduces PostgreSQL and MongoDB services, along with Mongo Express for easier database management. PostgreSQL will run on port 5432 while MongoDB is set to restart automatically and is accessible on port 27017. Additionally, Mongo Express is configured with admin credentials for easy access through port 10002. This is for CQRS lab